// src/interfaces/score.interface.ts
export interface ScoreType {
  opponent_1_name?: string
  opponent_1_energy?: number
  opponent_1_score?: number

  opponent_2_name?: string
  opponent_2_energy?: number
  opponent_2_score?: number

  energy_icon?: string
  location?: string
  month?: string,
  date?: number,
  year?: number
  short_date?: string
  time?: string
  sport_icon?: string
}